+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Fixed
 * Fix "can't alloc thread" error in Command#run from non-main thread's rescue/ensure block 
 
+### Changed
+* Enable color by default only when output is a TTY
+
 ## [v0.10.1] - 2021-02-14
 
 ### Fixed